requirements.txt
pytest==7.4.2
requests==2.31.0
PyYAML==6.0.1
pymysql==1.1.0
sqlalchemy==2.0.23
allure-pytest==2.13.2
pytest-cov==4.1.0
